+++ artifact_gen/lib/component/copy_with.dart
@@ -33,7 +33,7 @@ class $ArtifactCopyWithComponent implements $ArtifactBuilderOutput {
     List<Uri> importUris = <Uri>[];
     LibraryElement targetLib = clazz.library;
 
-    buf.writeln('  ${builder.applyDefsF(clazz.name)} copyWith({');
+    buf.write('  ${builder.applyDefsF(clazz.name)} copyWith({');
 
     for (ParameterElement param in params) {
       String name = param.name;
@@ -46,71 +46,81 @@ class $ArtifactCopyWithComponent implements $ArtifactBuilderOutput {
               : builder.applyDefsF(baseType);
       String forceNullType = "${builder.applyDefsF(baseType)}?";
 
-      buf.writeln('    $forceNullType $name,');
+      buf.write('    $forceNullType $name,');
       builder.registerDef("bool");
       if (param.isOptionalNamed && fullType.endsWith("?")) {
-        buf.writeln(
-          '    ${builder.applyDefsF("bool")} delete${name.capitalize()} = _F,',
+        buf.write(
+          '${builder.applyDefsF("bool")} delete${name.capitalize()}=_F,',
         );
       }
 
       if (param.hasDefaultValue) {
-        buf.writeln(
-          '    ${builder.applyDefsF("bool")} reset${name.capitalize()} = _F,',
+        buf.write(
+          '${builder.applyDefsF("bool")} reset${name.capitalize()}=_F,',
         );
       }
 
       Uri uri = builder.$getImport(type, targetLib);
       if (uri.toString().isNotEmpty) importUris.add(uri);
 
-      if (baseType == "int") {
-        buf.writeln(
-          '    ${builder.applyDefsF("int")}? delta${name.capitalize()},',
+      if (baseType.startsWith("List<") || baseType.startsWith("Set<")) {
+        buf.write(
+          '${builder.applyDefsF(baseType)}? append${name.capitalize()},',
+        );
+        buf.write(
+          '${builder.applyDefsF(baseType)}? remove${name.capitalize()},',
         );
       }
+
+      if (baseType == "int") {
+        buf.write('${builder.applyDefsF("int")}? delta${name.capitalize()},');
+      }
       if (baseType == "double") {
-        buf.writeln(
-          '    ${builder.applyDefsF("double")}? delta${name.capitalize()},',
+        buf.write(
+          '${builder.applyDefsF("double")}? delta${name.capitalize()},',
         );
       }
     }
-    buf.writeln('  }) '); // end of parameter list, start body
+    buf.write('})'); // end of parameter list, start body
 
-    buf.writeln('    => ${clazz.name}(');
+    buf.write('=>${builder.applyDefsF(clazz.name)}(');
     for (ParameterElement param in params) {
       String name = param.name;
       String bsn = param.type.getDisplayString(withNullability: true);
 
       String pref = "";
-      String nd = bsn.endsWith("?") ? " ?? 0" : "";
+      String nd = bsn.endsWith("?") ? "??0" : "";
       String bs = param.type.getDisplayString(withNullability: false);
       if (bs == "int") {
         pref =
-            "delta${name.capitalize()} != null ? ($name ?? _t.$name$nd) + delta${name.capitalize()} : ";
+            "delta${name.capitalize()}!=null?($name??_t.$name$nd)+delta${name.capitalize()}:";
       }
 
       if (bs == "double") {
         pref =
-            "delta${name.capitalize()} != null ? ($name ?? _t.$name$nd) + delta${name.capitalize()} : ";
+            "delta${name.capitalize()}!=null?($name??_t.$name$nd)+delta${name.capitalize()}:";
       }
 
+      String o = "";
+
       if (param.isOptionalNamed && bsn.endsWith("?") && param.hasDefaultValue) {
-        buf.writeln(
-          '      $name: $pref delete${name.capitalize()} ? null : reset${name.capitalize()} ? ${param.defaultValueCode} : ($name ?? _t.$name),',
-        );
+        o = '${pref}delete${name.capitalize()}?null:reset${name.capitalize()} ? ${builder.valD(param.defaultValueCode.toString(), param.type)}:($name??_t.$name),';
       } else if (param.isOptionalNamed && bsn.endsWith("?")) {
-        buf.writeln(
-          '      $name: $pref delete${name.capitalize()} ? null : ($name ?? _t.$name),',
-        );
+        o = '${pref}delete${name.capitalize()}?null:($name??_t.$name),';
       } else if (param.hasDefaultValue) {
-        buf.writeln(
-          '      $name: $pref reset${name.capitalize()} ? ${param.defaultValueCode} : ($name ?? _t.$name),',
-        );
+        o = '${pref}reset${name.capitalize()}?${builder.valD(param.defaultValueCode.toString(), param.type)}:($name??_t.$name),';
       } else {
-        buf.writeln('      $name: $pref$name ?? _t.$name,');
+        o = '$pref$name??_t.$name,';
       }
+
+      if (bs.startsWith("List<") || bs.startsWith("Set<")) {
+        o = "(${o.substring(0, o.length - 1)})?.\$u(append${name.capitalize()},remove${name.capitalize()}),";
+      }
+
+      buf.write('$name: $o');
     }
-    buf.writeln('    );');
+
+    buf.writeln(');');
 
     return (importUris, buf);
   }