+++ artifact_gen/lib/component/to_map.dart
@@ -1,6 +1,7 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:artifact_gen/builder.dart';
+import 'package:toxic/extensions/iterable.dart';
 
 class $ArtifactToMapComponent implements $ArtifactBuilderOutput {
   const $ArtifactToMapComponent();
@@ -31,22 +32,22 @@ class $ArtifactToMapComponent implements $ArtifactBuilderOutput {
     List<Uri> importUris = <Uri>[];
     LibraryElement targetLib = clazz.library;
     buf.writeln(
-      "  ${builder.applyDefsF("String")} toJson({bool pretty = false}) => ${builder.applyDefsF("ArtifactCodecUtil")}.j(pretty, toMap);",
+      "  ${builder.applyDefsF("String")} toJson({bool pretty=_F})=>${builder.applyDefsF("ArtifactCodecUtil")}.j(pretty, toMap);",
     );
-    buf.writeln("  ${builder.applyDefsF("Map<String, dynamic>")} toMap() {");
-    buf.writeln("    _;");
+    buf.write("  ${builder.applyDefsF("Map<String, dynamic>")} toMap(){");
+    buf.write("_;");
 
     List<String>? subs = ArtifactBuilder.$artifactSubclasses[clazz.name];
     if (subs != null && subs.isNotEmpty) {
       for (String sub in subs) {
-        buf.writeln('    if (_t is ${builder.applyDefsF(sub)}) {');
-        buf.writeln('      return (_t as ${builder.applyDefsF(sub)}).toMap();');
-        buf.writeln('    }');
+        buf.write('if (_t is ${builder.applyDefsF(sub)}){');
+        buf.write('return (_t as ${builder.applyDefsF(sub)}).toMap();');
+        buf.write('}');
       }
     }
 
-    buf.writeln(
-      '    return <${builder.applyDefsF("String")}, ${builder.applyDefsF("dynamic")}>{',
+    buf.write(
+      'return <${builder.applyDefsF("String")}, ${builder.applyDefsF("dynamic")}>{',
     );
 
     InterfaceType? supType = clazz.supertype;
@@ -56,8 +57,8 @@ class $ArtifactToMapComponent implements $ArtifactBuilderOutput {
         sup,
         throwOnUnresolved: false,
       )) {
-        buf.writeln(
-          "      ${builder.stringD('_subclass_${sup.name}')}: '${clazz.name}',",
+        buf.write(
+          "${builder.stringD('_subclass_${sup.name}')}: '${clazz.name}',",
         );
       }
       supType = supType.element.supertype;
@@ -73,12 +74,36 @@ class $ArtifactToMapComponent implements $ArtifactBuilderOutput {
         $ArtifactConvertMode.toMap,
       );
 
-      buf.writeln("      ${builder.stringD(name)}: ${conv.code},");
+      String? rn;
+      FieldElement? field = clazz.fields.select((i) => i.name == param.name);
+
+      if (field != null &&
+          ArtifactBuilder.$renameChecker.hasAnnotationOf(
+            field,
+            throwOnUnresolved: false,
+          )) {
+        rn =
+            ArtifactBuilder.$renameChecker
+                .firstAnnotationOf(field, throwOnUnresolved: false)!
+                .getField("newName")!
+                .toStringValue();
+      } else if (ArtifactBuilder.$renameChecker.hasAnnotationOf(
+        param,
+        throwOnUnresolved: false,
+      )) {
+        rn =
+            ArtifactBuilder.$renameChecker
+                .firstAnnotationOf(param, throwOnUnresolved: false)!
+                .getField("newName")!
+                .toStringValue();
+      }
+
+      buf.write("${builder.stringD(rn ?? name)}:${conv.code.trim()},");
       importUris.addAll(conv.imports);
     }
 
-    buf.writeln('    };');
-    buf.writeln('  }');
+    buf.write('};');
+    buf.writeln('}');
 
     return (importUris, buf);
   }