+++ artifact_gen/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.0.1
+* Support for artifact 1.0.7
+
 # 1.0.0
 
 * Initial Generator